refactor(af): backend factory for all 3 apps; extract Solr to submodule
Make the ActivityFinderBackend plugin manager the single factory for every
Activity Finder app (AF4, AF3, Camp Finder), not just AF4, and move Solr
out of the main module.

- Backend contract: add getCategoriesTopLevel(); move the shared constants
  (RESULTS_PER_PAGE, CACHE_TAG) onto OpenyActivityFinderBackendInterface so
  no consumer references the Solr class for them.
- All consumers resolve via the factory: AF3 ActivityFinderBlock,
  ActivityFinderSearchBlock, Camp Finder CampFinderBlock and the search_api
  processors now call createInstance($settings->get('backend')); SessionData
  fetches saved items through the aggregator with an id filter (items are
  results, not a bespoke getSessions); the aggregator stamps each row with its
  source backend id and dedupes per backend.
- SettingsForm backend select lists discovered plugins and stores a plugin id;
  the default backend setting is now 'mock'.
- Extract Solr into the openy_activity_finder_solr submodule: the Solr
  implementation, the 'solr' plugin and the 8 search_api processors move there
  with their own search_api + search_api_solr dependencies. The main module no
  longer contains or depends on Solr.
- hook_update_9006 enables the submodule and maps the legacy global service id
  (openy_activity_finder.solr_backend) to the 'solr' plugin, so existing sites
  keep Solr while fresh installs default to Mock.

see queue/W0b-backend-harness/ (factory across AF3/AF4/CF + Solr submodule)
