+++ fixtures/getCategoriesTopLevel.json
@@ -1,0 +1,8 @@
+[
+    "Kids and Family Activities",
+    "Health and Fitness",
+    "Swimming",
+    "Camps",
+    "Child Care",
+    "Youth Programs"
+]
