+++ src/Plugin/Block/ActivityFinderSearchBlock.php
@@ -4,7 +4,7 @@ namespace Drupal\openy_activity_finder\Plugin\Block;
 
 use Drupal\Core\Block\BlockBase;
 use Drupal\Core\Cache\Cache;
-use Drupal\openy_activity_finder\OpenyActivityFinderSolrBackend;
+use Drupal\openy_activity_finder\OpenyActivityFinderBackendInterface;
 
 /**
  * Provides a 'PEF Programs' block.
@@ -23,7 +23,7 @@ class ActivityFinderSearchBlock extends BlockBase {
   public function build() {
     $config = \Drupal::service('config.factory')->get('openy_activity_finder.settings');
     $backend_service_id = $config->get('backend');
-    $backend = \Drupal::service($backend_service_id);
+    $backend = \Drupal::service('plugin.manager.activity_finder_backend')->createInstance($backend_service_id);
 
     return [
       '#theme' => 'openy_activity_finder_program_search_page',
@@ -57,7 +57,7 @@ class ActivityFinderSearchBlock extends BlockBase {
    * {@inheritdoc}
    */
   public function getCacheTags() {
-    return Cache::mergeTags(parent::getCacheTags(), [OpenyActivityFinderSolrBackend::ACTIVITY_FINDER_CACHE_TAG]);
+    return Cache::mergeTags(parent::getCacheTags(), [OpenyActivityFinderBackendInterface::CACHE_TAG]);
   }
 
 }
